Restringir permisos del rol admin y forzarlos en el backend

+++ routes/web.php
@@ -63,31 +63,43 @@ Route::prefix('api')->name('api.')->group(function (): void {
         Route::put('user/profile', [CurrentUserController::class, 'updateProfile'])->name('user.profile.update');
         Route::put('user/password', [CurrentUserController::class, 'updatePassword'])->name('user.password.update');
 
-        Route::post('users/get-all', [UserController::class, 'getAll'])->name('users.get-all');
-        Route::put('users/{user}/reset-password', [UserController::class, 'resetPassword'])->name('users.reset-password');
-        Route::get('users/{user}/audits', [UserController::class, 'audits'])->name('users.audits');
-        Route::apiResource('users', UserController::class)->except(['index']);
+        Route::post('users/get-all', [UserController::class, 'getAll'])->name('users.get-all')->middleware('permission:users.view');
+        Route::put('users/{user}/reset-password', [UserController::class, 'resetPassword'])->name('users.reset-password')->middleware('permission:users.update');
+        Route::get('users/{user}/audits', [UserController::class, 'audits'])->name('users.audits')->middleware('permission:users.view');
+        Route::apiResource('users', UserController::class)->except(['index'])
+            ->middlewareFor('store', 'permission:users.create')
+            ->middlewareFor('show', 'permission:users.view')
+            ->middlewareFor(['update', 'put', 'patch'], 'permission:users.update')
+            ->middlewareFor('destroy', 'permission:users.delete');
 
-        Route::post('identification-types/get-all', [IdentificationTypeController::class, 'getAll'])->name('identification-types.get-all');
-        Route::apiResource('identification-types', IdentificationTypeController::class)->except(['index']);
+        Route::post('identification-types/get-all', [IdentificationTypeController::class, 'getAll'])->name('identification-types.get-all')->middleware('permission:identification-types.view');
+        Route::apiResource('identification-types', IdentificationTypeController::class)->except(['index'])
+            ->middlewareFor('store', 'permission:identification-types.create')
+            ->middlewareFor('show', 'permission:identification-types.view')
+            ->middlewareFor(['update', 'put', 'patch'], 'permission:identification-types.update')
+            ->middlewareFor('destroy', 'permission:identification-types.delete');
 
-        Route::post('roles/get-all', [RoleController::class, 'getAll'])->name('roles.get-all');
-        Route::apiResource('roles', RoleController::class)->except(['index']);
+        Route::post('roles/get-all', [RoleController::class, 'getAll'])->name('roles.get-all')->middleware('permission:roles.view');
+        Route::apiResource('roles', RoleController::class)->except(['index'])
+            ->middlewareFor('store', 'permission:roles.create')
+            ->middlewareFor('show', 'permission:roles.view')
+            ->middlewareFor(['update', 'put', 'patch'], 'permission:roles.update')
+            ->middlewareFor('destroy', 'permission:roles.delete');
 
-        Route::post('permissions/get-all', [PermissionController::class, 'getAll'])->name('permissions.get-all');
+        Route::post('permissions/get-all', [PermissionController::class, 'getAll'])->name('permissions.get-all')->middleware('permission:permissions.view');
 
-        Route::get('settings', [SettingController::class, 'index'])->name('settings.index');
-        Route::put('settings', [SettingController::class, 'update'])->name('settings.update');
+        Route::get('settings', [SettingController::class, 'index'])->name('settings.index')->middleware('permission:settings.view');
+        Route::put('settings', [SettingController::class, 'update'])->name('settings.update')->middleware('permission:settings.update');
 
-        Route::get('api-credentials', [ApiCredentialController::class, 'index'])->name('api-credentials.index');
-        Route::post('api-credentials', [ApiCredentialController::class, 'store'])->name('api-credentials.store');
-        Route::put('api-credentials/{id}', [ApiCredentialController::class, 'update'])->name('api-credentials.update');
-        Route::delete('api-credentials/{id}', [ApiCredentialController::class, 'destroy'])->name('api-credentials.destroy');
-        Route::post('api-credentials/{id}/regenerate', [ApiCredentialController::class, 'regenerate'])->name('api-credentials.regenerate');
-        Route::patch('api-credentials/{id}/toggle-status', [ApiCredentialController::class, 'toggleStatus'])->name('api-credentials.toggle-status');
-        Route::post('api-credentials/{id}/revoke', [ApiCredentialController::class, 'revoke'])->name('api-credentials.revoke');
-        Route::get('api-credentials/{id}/logs', [ApiCredentialController::class, 'logs'])->name('api-credentials.logs');
-        Route::get('api-credentials/{id}/stats', [ApiCredentialController::class, 'stats'])->name('api-credentials.stats');
+        Route::get('api-credentials', [ApiCredentialController::class, 'index'])->name('api-credentials.index')->middleware('permission:api-keys.view');
+        Route::post('api-credentials', [ApiCredentialController::class, 'store'])->name('api-credentials.store')->middleware('permission:api-keys.create');
+        Route::put('api-credentials/{id}', [ApiCredentialController::class, 'update'])->name('api-credentials.update')->middleware('permission:api-keys.update');
+        Route::delete('api-credentials/{id}', [ApiCredentialController::class, 'destroy'])->name('api-credentials.destroy')->middleware('permission:api-keys.delete');
+        Route::post('api-credentials/{id}/regenerate', [ApiCredentialController::class, 'regenerate'])->name('api-credentials.regenerate')->middleware('permission:api-keys.update');
+        Route::patch('api-credentials/{id}/toggle-status', [ApiCredentialController::class, 'toggleStatus'])->name('api-credentials.toggle-status')->middleware('permission:api-keys.update');
+        Route::post('api-credentials/{id}/revoke', [ApiCredentialController::class, 'revoke'])->name('api-credentials.revoke')->middleware('permission:api-keys.delete');
+        Route::get('api-credentials/{id}/logs', [ApiCredentialController::class, 'logs'])->name('api-credentials.logs')->middleware('permission:api-keys.view');
+        Route::get('api-credentials/{id}/stats', [ApiCredentialController::class, 'stats'])->name('api-credentials.stats')->middleware('permission:api-keys.view');
 
         Route::get('clinicas', [ClinicaController::class, 'index'])->name('clinicas.index');
         Route::post('clinicas', [ClinicaController::class, 'store'])->name('clinicas.store');
